isolate search and sidenav components
