+++ js/components/search.js
@@ -1,0 +1,3 @@
+$('.preview-search-bar .usa-search').submit(function (event){
+  event.preventDefault();
+});
